split tuvx source into separate headers and modules
