+++ fortran/tuvx/CMakeLists.txt
@@ -1,0 +1,8 @@
+target_sources(musica-fortran
+  PRIVATE
+    grid.F90
+    grid_map.F90
+    profile.F90
+    profile_map.F90
+    tuvx.F90
+)
