+++ src/tuvx/grid.cpp
@@ -1,0 +1,74 @@
+// Copyright (C) 2023-2024 National Center for Atmospheric Research
+// SPDX-License-Identifier: Apache-2.0
+//
+// This file contains the implementation of the TUVX class, which represents a multi-component
+// reactive transport model. It also includes functions for creating and deleting TUVX instances.
+#include <musica/tuvx/grid.hpp>
+
+#include <cstring>
+#include <filesystem>
+#include <iostream>
+
+namespace musica
+{
+
+  // Grid external C API functions
+
+  void DeleteGrid(Grid *grid, Error *error)
+  {
+    *error = NoError();
+    try
+    {
+      delete grid;
+    }
+    catch (const std::system_error &e)
+    {
+      *error = ToError(e);
+    }
+  }
+
+  void SetGridEdges(Grid *grid, double edges[], std::size_t num_edges, Error *error)
+  {
+    DeleteError(error);
+    grid->SetEdges(edges, num_edges, error);
+  }
+
+  void SetGridMidpoints(Grid *grid, double midpoints[], std::size_t num_midpoints, Error *error)
+  {
+    DeleteError(error);
+    grid->SetMidpoints(midpoints, num_midpoints, error);
+  }
+
+  // Grid class functions
+
+  Grid::~Grid()
+  {
+    int error_code = 0;
+    if (grid_ != nullptr)
+      InternalDeleteGrid(grid_, &error_code);
+    grid_ = nullptr;
+  }
+
+  void Grid::SetEdges(double edges[], std::size_t num_edges, Error *error)
+  {
+    int error_code = 0;
+    InternalSetEdges(grid_, edges, num_edges, &error_code);
+    *error = NoError();
+    if (error_code != 0)
+    {
+      *error = Error{ 1, CreateString(MUSICA_ERROR_CATEGORY), CreateString("Failed to set edges") };
+    }
+  }
+
+  void Grid::SetMidpoints(double midpoints[], std::size_t num_midpoints, Error *error)
+  {
+    int error_code = 0;
+    InternalSetMidpoints(grid_, midpoints, num_midpoints, &error_code);
+    *error = NoError();
+    if (error_code != 0)
+    {
+      *error = Error{ 1, CreateString(MUSICA_ERROR_CATEGORY), CreateString("Failed to set midpoints") };
+    }
+  }
+
+}  // namespace musica
